optimize is_bogon
